+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.js
@@ -12,6 +12,7 @@ import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { useResizeDetector } from 'react-resize-detector';
+import { moderate02 } from '@carbon/motion';
 import wrapFocus from '../../global/js/utils/wrapFocus';
 import { pkg } from '../../settings';
 import { allPropTypes } from '../../global/js/utils/props-helper';
@@ -341,7 +342,7 @@ export let SidePanel = React.forwardRef(
       }
     }, [open]);
 
-    // initialize the side panel to close
+    // initializes the side panel to close
     const onAnimationEnd = () => {
       if (!open) {
         onUnmount && onUnmount();
@@ -350,9 +351,13 @@ export let SidePanel = React.forwardRef(
       setAnimationComplete(true);
     };
 
-    // initializes the side panel to open and prevents the side panel from being scrolled during animation
-    const onAnimationStart = () => {
-      setAnimationComplete(false);
+    // initializes the side panel to open
+    const onAnimationStart = (event) => {
+      event.persist();
+      const isPanelTarget = event.target.id === `${blockClass}-outer`;
+      if (isPanelTarget) {
+        setAnimationComplete(false);
+      }
     };
 
     // used to reset margins of the slide in panel when closed/closing
@@ -377,11 +382,11 @@ export let SidePanel = React.forwardRef(
         );
         if (placement && placement === 'right' && pageContentElement) {
           pageContentElement.style.marginRight = 0;
-          pageContentElement.style.transition = 'margin-right 250ms';
+          pageContentElement.style.transition = `margin-right ${moderate02}`;
           pageContentElement.style.marginRight = SIDE_PANEL_SIZES[size];
         } else if (pageContentElement) {
           pageContentElement.style.marginLeft = 0;
-          pageContentElement.style.transition = 'margin-left 250ms';
+          pageContentElement.style.transition = `margin-left ${moderate02}`;
           pageContentElement.style.marginLeft = SIDE_PANEL_SIZES[size];
         }
       }
@@ -447,6 +452,7 @@ export let SidePanel = React.forwardRef(
           actionToolbarButtons && actionToolbarButtons.length,
         [`${blockClass}__container-without-overlay`]:
           !includeOverlay && !slideIn,
+        [`${blockClass}__container-is-animating`]: !animationComplete,
       },
     ]);
 
@@ -456,7 +462,8 @@ export let SidePanel = React.forwardRef(
           className={cx(`${blockClass}__title-container`, {
             [`${blockClass}__on-detail-step`]: currentStep > 0,
             [`${blockClass}__title-container--no-animation`]: !animateTitle,
-            [`${blockClass}__title-container-is-animating`]: !animationComplete,
+            [`${blockClass}__title-container-is-animating`]:
+              !animationComplete && animateTitle,
           })}>
           {currentStep > 0 && (
             <Button
@@ -492,7 +499,8 @@ export let SidePanel = React.forwardRef(
               [`${blockClass}__subtitle-text-no-animation-no-action-toolbar`]:
                 !animateTitle &&
                 (!actionToolbarButtons || !actionToolbarButtons.length),
-              [`${blockClass}__subtitle-text-is-animating`]: !animationComplete,
+              [`${blockClass}__subtitle-text-is-animating`]:
+                !animationComplete && animateTitle,
             })}>
             {subtitle}
           </p>
@@ -505,23 +513,31 @@ export let SidePanel = React.forwardRef(
             {actionToolbarButtons.map((action) => (
               <Button
                 key={action.label}
-                kind={action.leading ? action.kind : 'ghost'}
+                kind={action.kind || 'ghost'}
                 size="small"
-                disabled={false}
                 renderIcon={action.icon}
                 iconDescription={action.label}
                 tooltipPosition="bottom"
                 tooltipAlignment="center"
+                hasIconOnly={!action.leading}
+                disabled={action.disabled}
                 className={cx([
                   `${blockClass}__action-toolbar-button`,
+                  action.className,
                   {
                     [`${blockClass}__action-toolbar-icon-only-button`]:
-                      action.icon,
+                      action.icon && !action.leading,
                     [`${blockClass}__action-toolbar-leading-button`]:
-                      !action.icon,
+                      action.leading,
                   },
                 ])}
-                onClick={() => action.onActionToolbarButtonClick()}>
+                onClick={(event) =>
+                  action.onClick
+                    ? action.onClick(event)
+                    : action.onActionToolbarButtonClick
+                    ? action.onActionToolbarButtonClick(event)
+                    : null
+                }>
                 {action.leading && action.label}
               </Button>
             ))}
@@ -573,15 +589,15 @@ export let SidePanel = React.forwardRef(
               animation: `${
                 open
                   ? placement === 'right'
-                    ? 'sidePanelEntranceRight 250ms'
-                    : 'sidePanelEntranceLeft 250ms'
+                    ? `sidePanelEntranceRight ${moderate02}`
+                    : `sidePanelEntranceLeft ${moderate02}`
                   : placement === 'right'
-                  ? 'sidePanelExitRight 250ms'
-                  : 'sidePanelExitLeft 250ms'
+                  ? `sidePanelExitRight ${moderate02}`
+                  : `sidePanelExitLeft ${moderate02}`
               }`,
             }}
             onAnimationEnd={onAnimationEnd}
-            onAnimationStart={onAnimationStart}
+            onAnimationStart={(event) => onAnimationStart(event)}
             onBlur={handleBlur}
             ref={contentRef}
             role="complementary"
@@ -622,8 +638,8 @@ export let SidePanel = React.forwardRef(
               style={{
                 animation: `${
                   open
-                    ? 'sidePanelOverlayEntrance 250ms'
-                    : 'sidePanelOverlayExit 250ms'
+                    ? `sidePanelOverlayEntrance ${moderate02}`
+                    : `sidePanelOverlayExit ${moderate02}`
                 }`,
               }}
             />
@@ -669,7 +685,11 @@ SidePanel.propTypes = {
       label: PropTypes.string,
       leading: PropTypes.bool,
       icon: PropTypes.object,
-      onActionToolbarButtonClick: PropTypes.func,
+      onActionToolbarButtonClick: deprecateProp(
+        PropTypes.func,
+        'This prop will be removed in the future. Please use `onClick` instead'
+      ),
+      onClick: PropTypes.func,
       kind: PropTypes.oneOf(['ghost', 'tertiary', 'secondary', 'primary']),
     })
   ),