+++ packages/cloud-cognitive/src/global/js/hooks/index.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export { useWindowScroll } from './useWindowScroll';
+export { useNearestScroll, useWindowScroll } from './useWindowScroll';
 export { useWindowResize } from './useWindowResize';
 export { useClickOutside } from './useClickOutside';
 export { useCreateComponentFocus } from './useCreateComponentFocus';