+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,99 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.77.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.77.0...@carbon/ibm-cloud-cognitive@0.77.1) (2021-08-26)
+
+
+### Bug Fixes
+
+* **SidePanel:** fix animating state of panel ([#1193](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1193)) ([5a98abc](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/5a98abc93e06b021ea4851c19afc4bbb50a7660f))
+
+
+
+
+
+# [0.77.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.76.1...@carbon/ibm-cloud-cognitive@0.77.0) (2021-08-25)
+
+
+### Features
+
+* cascade component ([#1171](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1171)) ([76b8b6f](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/76b8b6f1b375e093e721ba415b53e97ec3ca2262))
+
+
+
+
+
+## [0.76.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.76.0...@carbon/ibm-cloud-cognitive@0.76.1) (2021-08-25)
+
+
+### Bug Fixes
+
+* better warnings when deprecated-usage props have invalid values ([#1198](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1198)) ([6ec8777](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/6ec8777b66b4f45d47ed1060051b8bf9ec8692d8))
+
+
+
+
+
+# [0.76.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.75.1...@carbon/ibm-cloud-cognitive@0.76.0) (2021-08-25)
+
+
+### Features
+
+* **toolbar:** add vertical variant ([#1173](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1173)) ([574a2c5](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/574a2c55fd965be4de163ea8d9d9fae0baf1e7ba))
+
+
+
+
+
+## [0.75.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.75.0...@carbon/ibm-cloud-cognitive@0.75.1) (2021-08-25)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+# [0.75.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.74.1...@carbon/ibm-cloud-cognitive@0.75.0) (2021-08-23)
+
+
+### Features
+
+* **SidePanel:** allow actionToolbarButtons to be a bit more flexible ([#1189](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1189)) ([bc84e20](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/bc84e20f9652ba0379b884e5a9e1c7542ef34601))
+
+
+
+
+
+## [0.74.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.74.0...@carbon/ibm-cloud-cognitive@0.74.1) (2021-08-23)
+
+
+### Bug Fixes
+
+* **SidePanel:** center close svg in close button ([#1190](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1190)) ([fa4a688](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/fa4a68898eb74d2e5487732d27f36664de9648a1))
+
+
+
+
+
+# [0.74.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.73.2...@carbon/ibm-cloud-cognitive@0.74.0) (2021-08-20)
+
+
+### Features
+
+* fix offset and add element based scroll ([#1185](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1185)) ([e3149ae](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/e3149aec52eec078eb3b77abd029131ad9365385))
+
+
+
+
+
+## [0.73.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.73.1...@carbon/ibm-cloud-cognitive@0.73.2) (2021-08-20)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
 ## [0.73.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.73.0...@carbon/ibm-cloud-cognitive@0.73.1) (2021-08-19)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive